# contracts/openapi/orchestrator-v0.1.json
{
  "components": {
    "schemas": {
...
      "ContactLookup": {
        "additionalProperties": false,
        "properties": {
          "contact_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Contact Id",
            "type": "string"
          }
        },
        "required": [
          "contact_id"
        ],
        "title": "ContactLookup",
        "type": "object"
      },
      "ContactMetadata": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "criticality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criticality"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "ephemeris_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ephemeris Id"
          },
          "ephemeris_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ephemeris Mode"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "mission_profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mission Profile Id"
          },
          "properties_cfes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Properties Cfes"
          },
          "properties_is_internal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Properties Is Internal"
          },
          "properties_is_test": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Properties Is Test"
          },
          "reservation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reservation Id"
          },
          "setup_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Setup Duration"
          },
          "signature_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Comment"
          },
          "signature_contains_human_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Contains Human Override"
          },
          "signature_last_impacting_principal_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Last Impacting Principal Kind"
          },
          "signature_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Outcome"
          },
          "spacecraft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spacecraft Id"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "station_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Station Id"
          },
          "system_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Id"
          },
          "teardown_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teardown Duration"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "title": "ContactMetadata",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "GatewayBearer": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "DART Orchestrator and API Proxy",
    "version": "0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v0/metadata/contact": {
      "post": {
        "description": "Accept a JSON contact lookup for dashboard form clients.",
        "operationId": "post_contact_metadata_v0_metadata_contact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactLookup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactMetadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bearer authentication failed."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The request is syntactically or semantically invalid."
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "An upstream dependency failed."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The service is not configured."
          }
        },
        "security": [
          {
            "GatewayBearer": []
          }
        ],
        "summary": "Post Contact Metadata"
      }
    },
    "/v0/metadata/contact/{contact_id}": {
      "get": {
        "description": "Return the stable projection of one KOGS contact.",
        "operationId": "contact_metadata_v0_metadata_contact__contact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "contact_id",
            "required": true,
            "schema": {
              "title": "Contact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactMetadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bearer authentication failed."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The request is syntactically or semantically invalid."
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "An upstream dependency failed."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "maxLength": 2000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "instance": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "urn:dart:problem:authentication",
                        "urn:dart:problem:configuration",
                        "urn:dart:problem:conflict",
                        "urn:dart:problem:domain",
                        "urn:dart:problem:internal",
                        "urn:dart:problem:method-not-allowed",
                        "urn:dart:problem:not-found",
                        "urn:dart:problem:upstream",
                        "urn:dart:problem:validation"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The service is not configured."
          }
        },
        "security": [
          {
            "GatewayBearer": []
          }
        ],
        "summary": "Contact Metadata"
      }
    },
...
  }
}